Fix merge; Move files
