+++ datamodel/openapi/openapi-generator/src/main/java/com/sap/cloud/sdk/datamodel/openapi/generator/customization/UseExcludePaths.java
@@ -1,8 +1,10 @@
-package com.sap.cloud.sdk.datamodel.openapi.generator;
+package com.sap.cloud.sdk.datamodel.openapi.generator.customization;
 
 import java.util.Objects;
 
 import javax.annotation.Nonnull;
+
+import com.sap.cloud.sdk.datamodel.openapi.generator.GeneratorCustomization;
 
 import io.swagger.v3.oas.models.OpenAPI;
 import lombok.Getter;
@@ -13,10 +15,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Getter
 @Slf4j
-public class GeneratorCustomizationUseExcludePaths
-    implements
-    GeneratorCustomization,
-    GeneratorCustomization.PreProcessOpenAPI
+public class UseExcludePaths implements GeneratorCustomization, GeneratorCustomization.PreProcessOpenAPI
 {
     private final String configKey = "excludePaths";
